Reorganisation des suites behat & ajout d'un ResourceContext spécifique à la création des ressources.
