+++ .travis.yml
@@ -25,25 +25,15 @@ install:
     - app/console cache:warmup --env=test
 
 before_script:
-
     - features/bootstrap/bootstrap.sh configure
     - features/bootstrap/bootstrap.sh test
 
 script:
   - phpunit
   - bin/behat -s installer
-  - bin/behat -s users
-  - bin/behat -s account
-  - bin/behat -s configs
-  - bin/behat -s games_admin
-  - bin/behat -s plugins_admin
-  - bin/behat -s users_admin
-  - bin/behat -s groups_admin
-  - bin/behat -s machines
-  - bin/behat -s minecraft
-  - bin/behat -s steam
-  - bin/behat -s teamspeak
-  - bin/behat -s teamspeak_instance
+  - bin/behat -s user
+  - bin/behat -s admin
+  - bin/behat -s servers
   - bin/behat -s dashboard
 
 after_script:
