+++ features/user_password.feature
@@ -1,4 +1,4 @@
-@account
+@user
 Feature: User account password change
   In order to enhance the security of my account
   As a logged user
