+++ src/DP/Core/CoreBundle/Behat/ServerContext.php
@@ -11,13 +11,6 @@
 
 namespace DP\Core\CoreBundle\Behat;
 
-use Behat\Behat\Hook\Scope\BeforeFeatureScope;
-use Behat\Gherkin\Node\TableNode;
-use DP\GameServer\MinecraftServerBundle\Entity\MinecraftServer;
-use DP\GameServer\SteamServerBundle\Entity\SteamServer;
-use DP\VoipServer\TeamspeakServerBundle\Entity\TeamspeakServer;
-use DP\VoipServer\TeamspeakServerBundle\Entity\TeamspeakServerInstance;
-
 class ServerContext extends DefaultContext
 {
     /**
@@ -200,239 +193,5 @@ class ServerContext extends DefaultContext
         } else {
             $this->assertSession()->elementsCount('css', sprintf('#%s.instance-list > .instance-item', $type), $amount);
         }
-    }
-
-    /**
-     * @Given /^there are following minecraft servers:$/
-     */
-    public function thereAreMinecraftServers(TableNode $table)
-    {
-        foreach ($table->getHash() as $data) {
-            $this->thereIsMinecraftServer(
-                $data['name'],
-                $data['machine'],
-                $data['port'],
-                $data['queryPort'],
-                $data['rconPort'],
-                $data['rconPassword'],
-                $data['game'],
-                $data['installDir'],
-                $data['maxplayers'],
-                $data['minHeap'],
-                $data['maxHeap'],
-                (isset($data['installed']) && $data['installed'] == 'yes'),
-                false
-            );
-        }
-
-        $this->getEntityManager()->flush();
-    }
-
-    public function thereIsMinecraftServer($name, $machine = null, $port = 25565, $queryPort = 25565, $rconPort = 25575, $rconPassword = 'test', $game = 'minecraft', $installDir = 'test', $maxplayers = 2, $minHeap = 128, $maxHeap = 256, $installed = true, $flush = true)
-    {
-        if (null === $server = $this->getRepository('minecraft')->findOneBy(array('name' => $name))) {
-            $game    = $this->thereIsGame($game);
-            $machine = $this->thereIsMachine($machine);
-
-            $server = new MinecraftServer();
-            $server->setName($name);
-            $server->setMachine($machine);
-            $server->setPort($port);
-            $server->setQueryPort($queryPort);
-            $server->setRconPort($rconPort);
-            $server->setRconPassword($rconPassword);
-            $server->setGame($game);
-            $server->setDir($installDir);
-            $server->setMaxplayers($maxplayers);
-            $server->setMinHeap($minHeap);
-            $server->setMaxHeap($maxHeap);
-
-            if ($installed) {
-                $server->setInstallationStatus(101);
-            }
-
-            $this->validate($server);
-
-            $this->getEntityManager()->persist($server);
-
-            if ($flush) {
-                $this->getEntityManager()->flush();
-            }
-        }
-
-        return $server;
-    }
-
-    /**
-     * @Given /^there are following steam servers:$/
-     */
-    public function thereAreSteamServers(TableNode $table)
-    {
-        foreach ($table->getHash() as $data) {
-            $this->thereIsSteamServer(
-                $data['name'],
-                $data['machine'],
-                $data['port'],
-                $data['rconPassword'],
-                $data['game'],
-                $data['installDir'],
-                $data['maxplayers'],
-                (isset($data['installed']) && $data['installed'] == 'yes'),
-                false
-            );
-        }
-
-        $this->getEntityManager()->flush();
-    }
-
-    public function thereIsSteamServer($name, $machine = null, $port = 27025, $rconPassword = 'test', $game = 'Counter-Strike', $installDir = 'test', $maxplayers = 2, $installed = true, $flush = true)
-    {
-        if (null === $server = $this->getRepository('steam')->findOneBy(array('name' => $name))) {
-            $game    = $this->thereIsGame($game);
-            $machine = $this->thereIsMachine($machine);
-
-            $server = new SteamServer();
-            $server->setName($name);
-            $server->setMachine($machine);
-            $server->setPort($port);
-            $server->setRconPassword($rconPassword);
-            $server->setGame($game);
-            $server->setDir($installDir);
-            $server->setMaxplayers($maxplayers);
-
-            if ($installed) {
-                $server->setInstallationStatus(101);
-            }
-
-            $this->validate($server);
-
-            $this->getEntityManager()->persist($server);
-
-            if ($flush) {
-                $this->getEntityManager()->flush();
-            }
-        }
-
-        return $server;
-    }
-
-    /**
-     * @Given /^there are following teamspeak servers:$/
-     */
-    public function thereAreTeamspeakServers(TableNode $table)
-    {
-        foreach ($table->getHash() as $data) {
-            $this->thereIsTeamspeakServer(
-                $data['machine'],
-                $data['queryPassword'],
-                $data['installDir'],
-                (isset($data['installed']) && $data['installed'] == 'yes'),
-                false
-            );
-        }
-
-        $this->getEntityManager()->flush();
-    }
-
-    public function thereIsTeamspeakServer($machine, $queryPassword = 'test', $installDir = 'test', $installed = true, $flush = true)
-    {
-        $machine = $this->thereIsMachine($machine);
-
-        if (null === $server = $this->getRepository('teamspeak')->findOneBy(array('machine' => $machine))) {
-            $server = new TeamspeakServer();
-            $server->setMachine($machine);
-            $server->setQueryPassword($queryPassword);
-            $server->setDir($installDir);
-
-            if ($installed) {
-                $server->setInstallationStatus(101);
-            }
-
-            $this->validate($server);
-
-            $this->getEntityManager()->persist($server);
-
-            if ($flush) {
-                $this->getEntityManager()->flush();
-            }
-        }
-
-        return $server;
-    }
-
-    /**
-     * @Given /^there are following teamspeak instances:$/
-     */
-    public function thereAreTeamspeakInstances(TableNode $table)
-    {
-        foreach ($table->getHash() as $data) {
-            $this->thereIsTeamspeakInstance(
-                $data['instanceId'],
-                $data['name'],
-                $data['server'],
-                $data['port'],
-                $data['slots'],
-                false
-            );
-        }
-
-        $this->getEntityManager()->flush();
-    }
-
-    public function thereIsTeamspeakInstance($instanceId, $name = 'Test', $server = 'test4@127.0.0.1', $port = 9887, $slots = 2, $flush = true)
-    {
-        $parts   = explode('@', $server);
-        $machine = $this->findOneBy('machine', array('username' => $parts[0]));
-        $server  = $this->findOneBy('teamspeak', array('machine' => $machine->getId()));
-
-        if (null === $instance = $this->getRepository('instance', 'teamspeak')->findOneBy(array('server' => $server->getId()))) {
-            $instance = new TeamspeakServerInstance($server);
-            $instance->setInstanceId($instanceId);
-            $instance->setName($name);
-            $instance->setPort($port);
-            $instance->setMaxClients($slots);
-            $instance->setAdminToken('test');
-            $instance->setAutostart(true);
-
-            $this->validate($server);
-
-            $this->getEntityManager()->persist($server);
-
-            if ($flush) {
-                $this->getEntityManager()->flush();
-            }
-        }
-
-        return $instance;
-    }
-
-    /**
-     * @param string $baseName
-     */
-    protected function getRepository($resource, $baseName = null)
-    {
-        $service = 'dedipanel.';
-
-        if (!empty($baseName)) {
-            $service .= $baseName . '.';
-        }
-
-        return $this->getService($service . 'repository.'.$resource);
-    }
-
-    protected function findOneBy($type, array $criteria, $repoPrefix = '')
-    {
-        $resource = $this
-            ->getRepository($type, $repoPrefix)
-            ->findOneBy($criteria)
-        ;
-
-        if (null === $resource) {
-            throw new \InvalidArgumentException(
-                sprintf('%s for criteria "%s" was not found.', str_replace('_', ' ', ucfirst($type)), serialize($criteria))
-            );
-        }
-
-        return $resource;
     }
 }
